fix(pi): treat a dead status owner PID as claimable
The pi-family status extension claims a pane by writing its PID into
ORCA_PI_STATUS_OWNED (ORCA_PRIME_AGENT_STATUS_OWNED for prime-agent), and any
process that sees a different PID there returns early. That keeps nested child
agents from double-reporting, which is what the guard is for.

It also never checked whether that owner is still running. A restart can hand the
new process tree its dead predecessor's PID, so every later load takes the early
return, no handlers register, and nothing reaches the agent-hooks endpoint. The
pane's status stays blank until someone clears the env var by hand.

The guard now probes liveness before suppressing. A live foreign owner still
suppresses, so the anti-double-report behavior is unchanged. EPERM counts as
alive, since it means the process exists but belongs to another user, and a
malformed PID is treated as dead rather than trusted.

Testing: the extension harness had no `process.kill`, so it gained an injectable
one. Tests declare which PIDs are live instead of probing the runner's real
process table, which keeps them deterministic. Reverting the guard fails the five
new reclaim tests; the EPERM case passes either way on purpose, guarding the
suppression path.

The harness moved to its own module because these tests pushed the file past the
800-line test budget. It is a straight move, no behavior change, matching the
existing *-test-harness.ts pattern.

Fixes #16109
