+++ src/main/pi/agent-status-handler-source.ts
@@ -58,13 +58,27 @@ export function getPiAgentStatusHandlerSourceLines(kind: PiAgentKind): string[] 
     '// etc.), so we forward the raw object verbatim under the same field',
     '// names Claude uses (tool_name / tool_input) and let the server pick the',
     '// preview. Keeps tool-name knowledge centralized on the receiver side.',
+    "// Why: a restarted agent can inherit a dead owner's PID, which would suppress",
+    '// its status reporting forever. Treat a dead owner as claimable; EPERM means the',
+    '// process exists under another user, so it still counts as alive.',
+    'function isOwnerAlive(pid: string): boolean {',
+    '  const parsed = Number(pid)',
+    '  if (!Number.isInteger(parsed) || parsed < 1) return false',
+    '  try {',
+    '    process.kill(parsed, 0)',
+    '    return true',
+    '  } catch (error) {',
+    "    return (error as { code?: string }).code === 'EPERM'",
+    '  }',
+    '}',
+    '',
     "// Why: child agents inherit the lead's pane env; only its process may",
     '// register status hooks. PID identity keeps in-process reloads reporting.',
     'export default function (pi): void {',
     ...primeDaemonWorkerGuard,
     `  const ownerPid = process.env.${ownerEnv}`,
     '  const selfPid = String(process.pid)',
-    '  if (ownerPid && ownerPid !== selfPid) return',
+    '  if (ownerPid && ownerPid !== selfPid && isOwnerAlive(ownerPid)) return',
     `  process.env.${ownerEnv} = selfPid`,
     ...sessionStartHandler,
     `  pi.on('before_agent_start', (event${ctxParam}) => {`,
